Move nuke module to its own package
